Simplify Unown form handling
"pokemon_forms.c" has been added which will eventually be used for the handling of Pokémon forms.

The handling of getting Unown's form through its personality has been put in GetUnownSpecies in pokemon_forms.c which gets the value of GetUnownLetterByPersonality and uses that to access Unown's new form table to get the respective Unown form species.

Anywhere that does the respective checks for Unown have been replaced with a call to GetUnownSpecies.