Implement "relativeFrames" for sprites
A new field has been added to "struct SpriteFrameImage" which when enabled allow sprites to be loaded correctly if the frames are in order. By adding this, arrays using "struct SpriteFrameImage" will only need one entry which is a straight 8 bytes as opposed to 8 bytes * the number of frames.

Applicable places have been changed to use this updated functionality.